fix(control): master self-entry identity + peer-IP pin matching; bump v0.2.12-rc.2
Two rc.1 prod findings from the placement exact-matcher rollout:

- #135: agents self-report hostname:port addresses, so bare-IP pins
  matched nothing and deploys refused. The master now records each
  control session's peer IP at the WS upgrade (ConnectInfo) on
  RegisteredNode.peer_ip and node_matches_pin matches against it — IP
  pins work regardless of the agent's self-reported address.
- #134: the master self-registers in registered_nodes; the entry was
  'localhost:{api_port}' with no hostname label, so a localhost pin
  resolved to it and attempted a WS dispatch to a session the master
  doesn't have. The self-entry now carries the real hostname (address
  + hostname label), and find_target_node deploys locally for any pin
  resolving to a role=master entry.

Serve paths now use into_make_service_with_connect_info so the WS
handler can extract the peer address (main server + WS test
harnesses). Bump workspace to 0.2.12-rc.2; changelog for both.

Fixes #134
Fixes #135
