+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- **Placement pins match the control session's peer IP (rc.2).** Agents self-report a `hostname:port` address, so under exact matching a bare-IP pin (`node = "217.154.26.121"`) matched nothing and deploys refused. The master now records each session's peer IP as seen at the WebSocket upgrade and matches pins against it — IP pins work regardless of what the agent reports. (#135)
+- **Master's self-entry no longer hijacks master pins into a WS dispatch (rc.2).** The master self-registers in the node map; it previously did so as `localhost:{port}`, so a `localhost` pin resolved to that entry and tried a remote dispatch to a session the master doesn't have. The self-entry now carries the master's real hostname identity, and any pin resolving to a `role=master` entry deploys locally. (#134)
 - **Half-dead agent sessions can no longer masquerade as healthy nodes.** The master↔agent control channel now enforces read-idle deadlines on both ends (master: `[deploy].ws_idle_timeout_secs`, default 30s; agent: 90s), so a half-open TCP connection — NAT timeout, VM freeze — tears the session down instead of persisting invisibly: the node is marked unreachable, its stale service state is dropped rather than reported as `Running`, and the agent reconnects on its own with backoff. A missed deploy ACK also kills the session immediately, so subsequent deploys fail fast with the real story instead of re-timing-out against a dead channel; and a reconnect can no longer be clobbered by its superseded session's late cleanup. Previously a dead channel could persist for days while heartbeats kept the node looking green and `orca status` served hours-stale state. (#131)
 - **Placement pins match exactly — no more substring split-brain.** `node = "ubuntu"` no longer matches a node named `ubuntu-16gb-fsn1-1`; pins match only on exact node id, full `ip:port` address, the address's host portion (bare-IP pins keep working), or the hostname label. A pin matching multiple nodes refuses to schedule (loud error) instead of picking a HashMap-ordered winner; a pin matching nothing (or only drained nodes) fails the deploy with a precise error instead of silently deploying to the master; and deploy targeting and reconcile now share one resolver so agents are never told to run what deploy refused. Explicit master pins (`master`, `localhost`, `127.0.0.1`, the master's hostname) still deploy locally. (#124)
 - **The proxy no longer forces HTTP→HTTPS redirects when no TLS endpoint exists.** With ACME unconfigured, requests for routed hosts were 301'd to an `https://` that nothing served; the redirect now fires only when a TLS listener or ACME manager is actually present. This also unblocks running orca behind an external TLS-terminating proxy (NGINX/Caddy), which previously redirect-looped; fuller behind-proxy support (forwarded headers, per-client rate limiting) is tracked in #125. (#123)
