+++ crates/orca-cli/Cargo.toml
@@ -14,10 +14,10 @@ path = "src/main.rs"
 [dependencies]
 orca-core = { workspace = true }
 orca-ai = { workspace = true }
-orca-agent = { path = "../orca-agent", version = "0.2.12-rc.1" }
-orca-control = { path = "../orca-control", version = "0.2.12-rc.1" }
+orca-agent = { path = "../orca-agent", version = "0.2.12-rc.2" }
+orca-control = { path = "../orca-control", version = "0.2.12-rc.2" }
 orca-proxy = { workspace = true }
-orca-tui = { path = "../orca-tui", version = "0.2.12-rc.1" }
+orca-tui = { path = "../orca-tui", version = "0.2.12-rc.2" }
 tokio = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
