+++ crates/orca-control/tests/failure_reasons_test.rs
@@ -169,7 +169,14 @@ async fn crash_heartbeat_records_failure() {
         .with_state(state.clone());
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
-    tokio::spawn(async move { axum::serve(listener, app).await.unwrap() });
+    tokio::spawn(async move {
+        axum::serve(
+            listener,
+            app.into_make_service_with_connect_info::<std::net::SocketAddr>(),
+        )
+        .await
+        .unwrap()
+    });
     tokio::time::sleep(Duration::from_millis(50)).await;
 
     let url = format!("ws://{addr}/api/v1/ws/agent?token=tok&node_id=7");
