+++ crates/orca-control/tests/remote_state_test.rs
@@ -162,7 +162,14 @@ async fn deploy_result_updates_remote_instance_to_running() {
         .with_state(state.clone());
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
-    tokio::spawn(async move { axum::serve(listener, app).await.unwrap() });
+    tokio::spawn(async move {
+        axum::serve(
+            listener,
+            app.into_make_service_with_connect_info::<std::net::SocketAddr>(),
+        )
+        .await
+        .unwrap()
+    });
     tokio::time::sleep(Duration::from_millis(50)).await;
 
     let url = format!("ws://{addr}/api/v1/ws/agent?token=test-tok&node_id=7");
