+++ crates/orca-control/tests/ws_integration_test.rs
@@ -55,7 +55,12 @@ async fn start_server(state: Arc<orca_control::state::AppState>) -> std::net::So
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
     tokio::spawn(async move {
-        axum::serve(listener, app).await.unwrap();
+        axum::serve(
+            listener,
+            app.into_make_service_with_connect_info::<std::net::SocketAddr>(),
+        )
+        .await
+        .unwrap();
     });
     // Give server a moment to start
     tokio::time::sleep(Duration::from_millis(50)).await;
@@ -109,6 +114,7 @@ async fn ws_heartbeat_updates_node_stats() {
         nodes.insert(
             42,
             orca_control::state::RegisteredNode {
+                peer_ip: None,
                 node_id: 42,
                 address: "test:6881".into(),
                 labels: HashMap::new(),
