+++ crates/orca-control/tests/ws_reconcile_test.rs
@@ -38,7 +38,12 @@ async fn start_server(state: Arc<orca_control::state::AppState>) -> std::net::So
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
     tokio::spawn(async move {
-        axum::serve(listener, app).await.unwrap();
+        axum::serve(
+            listener,
+            app.into_make_service_with_connect_info::<std::net::SocketAddr>(),
+        )
+        .await
+        .unwrap();
     });
     tokio::time::sleep(Duration::from_millis(50)).await;
     addr
@@ -54,6 +59,7 @@ async fn ws_sends_reconcile_on_connect() {
         nodes.insert(
             77,
             orca_control::state::RegisteredNode {
+                peer_ip: None,
                 node_id: 77,
                 address: "contabo-host:6881".into(),
                 labels: HashMap::new(),
